refactor: Move historyReducer to reducer folder
