+++ src/contexts/historyContext.js
@@ -1,20 +1,9 @@
 import axios from "axios";
 import { createContext, useReducer, useContext, useEffect } from "react";
 import { useAuth } from "./authContext";
+import { historyReducer } from "../reducer/historyReducer";
 
 const HistoryContext = createContext();
-
-const historyReducer = (state, action) => {
-	switch (action.type) {
-	case "SET_HISTORY_LIST":
-		return {
-			...state,
-			historyItems: action.payload,
-		};
-		default:
-			return state;
-	}
-};
 
 const HistoryProvider = ({ children }) => {
